Move SimilarCustomersTable queries and mutations next to it
