+++ src/features/applicationView/similarCustomersTable/mutations.ts
